Fix: MetaArticle::default_pub_date()
This now returns the current date and not the Date::Min Constant

This means the
```rust
    fn is_default_pub_date(&self) -> bool {
        self.pub_date == Date::MIN
    }
```
Is probably no longer requested.

Fix issue naming when creating a new Zine

Fix Author name being lowercase

Fix: Remove set_intro

From the code it appears the intro come from a seperate intro.md file

Fix: Remove setting slug in finalize

The slug should not be set as part of finalize at this time.
It might be good to be able to set this to /year/month/ in the future
based on a global option.

Fix: Typo

Fix: Issues regarding finalize MetaArticle

Fixed issues related to finalising the Article Struct.

There is still a bug if the title ends with a period (.) as this would
mean the generated file name would be "title..md". This should be
handled at some point.