Fix: Added code to handle duplicate issue creation
There was a case that miight only happen if a use did not use the next
available issue number.
If issue X existed and the user tried to provided the same value for X.
The program would report an error(17) file exists error.
This code should catch that and exit cleanly with a warning.